feat(component-events): replace components-core with unity-react-core
BREAKING CHANGE: Replace components-core with unity-react-core
