+++ packages/component-events/jest.config.js
@@ -9,6 +9,6 @@ module.exports = {
     "\\.(jpg|jpeg|png|gif|webp|svg)$": "jest-transform-file",
   },
   moduleNameMapper: {
-    "^@asu/components-core$": "<rootDir>/../components-core/src/index.js",
+    "^@asu/unity-react-core$": "<rootDir>/../unity-react-core/src/index.js",
   },
 };
