+++ packages/component-events/package.json
@@ -25,7 +25,7 @@
   },
   "scripts": {
     "lint": "eslint --fix 'src/**/*.{js,jsx}' --ignore-path ../../.eslintignore",
-    "test": "jest --config=./jest.config.js --passWithNoTests --silent --coverage",
+    "test": "vitest --watch=false",
     "test-update-snapshot": "yarn test -- -u",
     "build": "vite build && yarn postbuild",
     "build:stats": "webpack -c webpack/webpack.prod.js --profile --json=compilation-stats.json",
@@ -39,7 +39,7 @@
     "postdocs": "node ../../scripts/process-readme-props.js"
   },
   "dependencies": {
-    "@asu/components-core": "^3.0.0",
+    "@asu/unity-react-core": "workspace:^",
     "prop-types": "^15.7.2",
     "styled-components": "^5.3.0"
   },
@@ -52,7 +52,7 @@
     "@storybook/addon-links": "^7.6.14",
     "@storybook/react": "^7.6.14",
     "@storybook/react-webpack5": "^7.6.14",
-    "@testing-library/jest-dom": "^5.14.1",
+    "@testing-library/jest-dom": "^6.6.3",
     "@testing-library/react": "^16.0.0",
     "babel-loader": "^8.2.2",
     "copy-webpack-plugin": "^9.0.1",
@@ -83,6 +83,7 @@
     "storybook-addon-mock": "^4.3.0",
     "style-loader": "^2.0.0",
     "terser-webpack-plugin": "^5.1.1",
+    "vitest": "^3.0.5",
     "webpack-filter-warnings-plugin": "^1.2.1",
     "webpack-merge": "^5.8.0"
   },
