+++ packages/component-events/src/components/CardsGridEvents/index.jsx
